+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changelog
 
 - Fix non-WAV master audio, allow disabling "Preserve Mono Volume" (#517)
+- Replace QMutex with threading.Lock to avoid memory corruption issues (#524)
 
 ## 0.11.0